pyproject.toml
[tool.poetry]
name = "theregiment"
version = "0.1.0"
description = "MonsterCoach 2025 - AI-powered coaching platform"
authors = ["TheRegiment Team <team@theregiment.ai>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.5"
pydantic = "2.5.0"
uvicorn = "0.22.0"
python-dotenv = "^1.0.1"
"discord.py" = "^2.3.2"
asyncpg = "^0.29.0"
dspy-ai = "2.3.4"
apscheduler = "3.10.4"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
"pytest-asyncio" = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 